# nettest/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "App for analyze networks payload",
        "title": "AnalyzeNetworkApp",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/analysis-requests": {
            "get": {
                "description": "Retrieves a list of analysis requests based on the provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisRequests"
                ],
                "summary": "Get list of analysis requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis request status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date in the format '2006-01-02T15:04:05Z'",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date in the format '2006-01-02T15:04:05Z'",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AnalysisRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/analysis-requests/client": {
            "put": {
                "description": "Updates the status of an analysis request by client on registered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisRequests"
                ],
                "summary": "Update analysis request status by client",
                "parameters": [
                    {
                        "description": "New status of the analysis request",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/analysis-requests/{id}": {
            "get": {
                "description": "Retrieves an analysis request with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisRequests"
                ],
                "summary": "Get analysis request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Analysis Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/analysis-requests/{requestId}": {
            "delete": {
                "description": "Deletes an analysis request for the given user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisRequests"
                ],
                "summary": "Delete analysis request by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/analysis-requests/{requestId}/admin": {
            "put": {
                "description": "Updates the status of an analysis request with the given ID on \"COMPLETE\"/\"CANCELED\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisRequests"
                ],
                "summary": "Update analysis request status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New request status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/check-auth": {
            "get": {
                "description": "Retrieves user information based on the provided user context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Check user authentication",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "description": "Logs out the user by blacklisting the access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AnalysisRequest": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "string"
                },
                "adminId": {
                    "type": "integer"
                },
                "completeDate": {
                    "type": "string"
                },
                "creationDate": {
                    "type": "string"
                },
                "formationDate": {
                    "type": "string"
                },
                "requestId": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "isAdmin": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}